backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.12.0
pypdf==4.0.1
jinja2==3.1.3
weasyprint==60.2
sqlalchemy==2.0.27